+++ changelog.md
@@ -70,6 +70,8 @@
     - Triggers can be set to not deactivate automatically, making them activate only once until the room resets
     - Triggers can be set to have an activation period, requiring something to stay in the trigger for a certain amount of time to actually set the bound flag; same for deactivating
     - Progression Triggers grant a progression entry and optionally give an item as well once entered by a player; if the progression entry is already unlocked//obtained, it won't give an item
+    - Redstone Receivers set a flag when their redstone power state changes
+    - Timer Triggers start or stop a Players Timer; only works inside Levels.
   - Listener Blocks
     - Listen for the state of the bound Flag being changed and performs an action based on its type
     - All Listeners can have a set delay for their action; if the bound flag is deactivated again before the activation cooldown is run out, it'll be cancelled.